# openapi.yaml
openapi: 3.0.0
info:
  title: SPK
  version: 1.0.0
  description: ini untuk spk
servers:
  - url: '{{BASE_URL}}'
  - url: '{{PROCESSOR_URL}}'
paths:
  /register:
    post:
      summary: register
      responses:
        '200':
          description: '200'
          headers:
            Content-Security-Policy:
              schema:
                type: string
              example: >-
                default-src 'self';base-uri 'self';font-src 'self' https:
                data:;form-action 'self';frame-ancestors 'self';img-src 'self'
                data:;object-src 'none';script-src 'self';script-src-attr
                'none';style-src 'self' https:
                'unsafe-inline';upgrade-insecure-requests
            Cross-Origin-Opener-Policy:
              schema:
                type: string
              example: same-origin
            Origin-Agent-Cluster:
              schema:
                type: string
              example: '?1'
            Referrer-Policy:
              schema:
                type: string
              example: no-referrer
            Strict-Transport-Security:
              schema:
                type: string
              example: max-age=31536000; includeSubDomains
            X-Content-Type-Options:
              schema:
                type: string
              example: nosniff
            X-DNS-Prefetch-Control:
              schema:
                type: string
              example: 'off'
            X-Download-Options:
              schema:
                type: string
              example: noopen
            X-Frame-Options:
              schema:
                type: string
              example: SAMEORIGIN
            X-Permitted-Cross-Domain-Policies:
              schema:
                type: string
              example: none
            X-XSS-Protection:
              schema:
                type: integer
              example: 0
            Vary:
              schema:
                type: string
              example: Origin
            Access-Control-Allow-Credentials:
              schema:
                type: boolean
              example: true
            Content-Length:
              schema:
                type: integer
              example: 28
            ETag:
              schema:
                type: string
              example: W/"1c-87CmT96XRfMUd4Xz7iVy7wA+Vl0"
            Date:
              schema:
                type: string
              example: Thu, 14 May 2026 19:04:26 GMT
            Connection:
              schema:
                type: string
              example: keep-alive
            Keep-Alive:
              schema:
                type: string
              example: timeout=5
        '400':
          description: '400'
          headers:
            Content-Security-Policy:
              schema:
                type: string
              example: >-
                default-src 'self';base-uri 'self';font-src 'self' https:
                data:;form-action 'self';frame-ancestors 'self';img-src 'self'
                data:;object-src 'none';script-src 'self';script-src-attr
                'none';style-src 'self' https:
                'unsafe-inline';upgrade-insecure-requests
            Cross-Origin-Opener-Policy:
              schema:
                type: string
              example: same-origin
            Origin-Agent-Cluster:
              schema:
                type: string
              example: '?1'
            Referrer-Policy:
              schema:
                type: string
              example: no-referrer
            Strict-Transport-Security:
              schema:
                type: string
              example: max-age=31536000; includeSubDomains
            X-Content-Type-Options:
              schema:
                type: string
              example: nosniff
            X-DNS-Prefetch-Control:
              schema:
                type: string
              example: 'off'
            X-Download-Options:
              schema:
                type: string
              example: noopen
            X-Frame-Options:
              schema:
                type: string
              example: SAMEORIGIN
            X-Permitted-Cross-Domain-Policies:
              schema:
                type: string
              example: none
            X-XSS-Protection:
              schema:
                type: integer
              example: 0
            Vary:
              schema:
                type: string
              example: Origin
            Access-Control-Allow-Credentials:
              schema:
                type: boolean
              example: true
            Content-Length:
              schema:
                type: integer
              example: 71
            ETag:
              schema:
                type: string
              example: W/"47-AxbBiD9nw2yUPfnQY+eLIaH5Uzo"
            Date:
              schema:
                type: string
              example: Thu, 14 May 2026 19:15:08 GMT
            Connection:
              schema:
                type: string
              example: keep-alive
            Keep-Alive:
              schema:
                type: string
              example: timeout=5
      tags:
        - Auth
      requestBody:
        content:
          application/json:
            schema:
              type: object
            example: |-
              {
                  "nama": "rafi",
                  // "role":"User", // opsional
                  "username": "username",
                  "password": "password"
              }
  /login:
    post:
      summary: login
      responses:
        '200':
          description: '200'
          headers:
            Content-Security-Policy:
              schema:
                type: string
              example: >-
                default-src 'self';base-uri 'self';font-src 'self' https:
                data:;form-action 'self';frame-ancestors 'self';img-src 'self'
                data:;object-src 'none';script-src 'self';script-src-attr
                'none';style-src 'self' https:
                'unsafe-inline';upgrade-insecure-requests
            Cross-Origin-Opener-Policy:
              schema:
                type: string
              example: same-origin
            Origin-Agent-Cluster:
              schema:
                type: string
              example: '?1'
            Referrer-Policy:
              schema:
                type: string
              example: no-referrer
            Strict-Transport-Security:
              schema:
                type: string
              example: max-age=31536000; includeSubDomains
            X-Content-Type-Options:
              schema:
                type: string
              example: nosniff
            X-DNS-Prefetch-Control:
              schema:
                type: string
              example: 'off'
            X-Download-Options:
              schema:
                type: string
              example: noopen
            X-Frame-Options:
              schema:
                type: string
              example: SAMEORIGIN
            X-Permitted-Cross-Domain-Policies:
              schema:
                type: string
              example: none
            X-XSS-Protection:
              schema:
                type: integer
              example: 0
            Vary:
              schema:
                type: string
              example: Origin
            Access-Control-Allow-Credentials:
              schema:
                type: boolean
              example: true
            Set-Cookie:
              schema:
                type: string
              example: >-
                access_token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOjEsImlhdCI6MTc3ODc4NjA4MywiZXhwIjoxNzc4Nzg2OTgzfQ.IcYKjrvnaH1wFq1jYKH7tFT7vd5RlLanYJ_a3I1ZnOY;
                Max-Age=900; Path=/; Expires=Thu, 14 May 2026 19:29:43 GMT;
                HttpOnly; SameSite=Lax
            Content-Length:
              schema:
                type: integer
              example: 191
            ETag:
              schema:
                type: string
              example: W/"bf-DWjL1nujiYrwXHNGs+Wzlb3h/LY"
            Date:
              schema:
                type: string
              example: Thu, 14 May 2026 19:14:43 GMT
            Connection:
              schema:
                type: string
              example: keep-alive
            Keep-Alive:
              schema:
                type: string
              example: timeout=5
        '400':
          description: '400'
          headers:
            Content-Security-Policy:
              schema:
                type: string
              example: >-
                default-src 'self';base-uri 'self';font-src 'self' https:
                data:;form-action 'self';frame-ancestors 'self';img-src 'self'
                data:;object-src 'none';script-src 'self';script-src-attr
                'none';style-src 'self' https:
                'unsafe-inline';upgrade-insecure-requests
            Cross-Origin-Opener-Policy:
              schema:
                type: string
              example: same-origin
            Origin-Agent-Cluster:
              schema:
                type: string
              example: '?1'
            Referrer-Policy:
              schema:
                type: string
              example: no-referrer
            Strict-Transport-Security:
              schema:
                type: string
              example: max-age=31536000; includeSubDomains
            X-Content-Type-Options:
              schema:
                type: string
              example: nosniff
            X-DNS-Prefetch-Control:
              schema:
                type: string
              example: 'off'
            X-Download-Options:
              schema:
                type: string
              example: noopen
            X-Frame-Options:
              schema:
                type: string
              example: SAMEORIGIN
            X-Permitted-Cross-Domain-Policies:
              schema:
                type: string
              example: none
            X-XSS-Protection:
              schema:
                type: integer
              example: 0
            Vary:
              schema:
                type: string
              example: Origin
            Access-Control-Allow-Credentials:
              schema:
                type: boolean
              example: true
            Content-Length:
              schema:
                type: integer
              example: 72
            ETag:
              schema:
                type: string
              example: W/"48-foe7IEVlwZtXi7MWk6cAGyIexo4"
            Date:
              schema:
                type: string
              example: Thu, 14 May 2026 19:16:13 GMT
            Connection:
              schema:
                type: string
              example: keep-alive
            Keep-Alive:
              schema:
                type: string
              example: timeout=5
      tags:
        - Auth
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
            example:
              username: username
              password: password
...